refactor the common code to change the way that descriptions are represented in the WB.
Now we use FSN / SYNONYM - and put the term definition type in as an annotation.

Likewise with relationships - but IS_A is the only one we have remapped so far.

Big API cleanups when along with this change to the common code.

Other minor bug fixes to the NDF loader (moved generic to description from attribute) and VHAT (got rid of dupe refsets)